Factor member name lookup out of `handle_name.cpp`.
